moved chartvotepower to staking
