+++ app/src/renderer/components/staking/ChartVotePower.vue
